requirements.txt
Flask
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai>=0.3.0
python-dotenv
woocommerce
fuzzywuzzy
python-Levenshtein
Pillow